requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Data handling
h5py>=3.7.0
pandas>=1.3.0

# Training and logging
tensorboard>=2.10.0
tqdm>=4.62.0
pyyaml>=6.0
wandb>=0.16.0

# Evaluation and metrics
edit-distance>=1.0.4
matplotlib>=3.5.0
seaborn>=0.11.0
numba>=0.55.0
